source/requirements.txt
python-telegram-bot==13.11
requests==2.24.0
python-dotenv==0.15.0